Relocate MaskSecret comms tests into task SDK comms tests
